# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/awaiting-confirm-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all awaiting confirm request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "awaiting confirm request",
                    "admin"
                ],
                "summary": "Get all awaiting confirm request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageNumber",
                        "name": "page[number]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "page[size]",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "symbols",
                        "name": "symbols",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "flowTypes",
                        "name": "flowTypes",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "accountNumbers",
                        "name": "accountNumbers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "investorId",
                        "name": "investorId",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ids",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "startDate",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endDate",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "latestUpdateCategories",
                        "name": "latestUpdateCategories",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "assetType",
                        "name": "assetType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "custodyCode",
                        "name": "custodyCode",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "custodyCodes",
                        "name": "custodyCodes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ResponseWithPaging-array_financing-offer_internal_core_entity_AwaitingConfirmRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/derivative-awaiting-confirm-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all awaiting confirm request derivative",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "awaiting confirm request",
                    "admin"
                ],
                "summary": "Get all awaiting confirm request derivative",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageNumber",
                        "name": "page[number]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "page[size]",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "symbols",
                        "name": "symbols",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "flowTypes",
                        "name": "flowTypes",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "accountNumbers",
                        "name": "accountNumbers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "investorId",
                        "name": "investorId",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ids",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "startDate",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endDate",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "latestUpdateCategories",
                        "name": "latestUpdateCategories",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "assetType",
                        "name": "assetType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "custodyCode",
                        "name": "custodyCode",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "custodyCodes",
                        "name": "custodyCodes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ResponseWithPaging-array_financing-offer_internal_core_entity_AwaitingConfirmRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/investor-accounts/{account-no}/margin-status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verify and update investor_account margin status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investor account margin status",
                    "admin"
                ],
                "summary": "Verify and update investor_account margin status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account-no",
                        "name": "accountNo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_core_investor_account_transport_http.VerifyInvestorAccountMarginStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_InvestorAccount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "financing-offer_internal_core.PagingMetaData": {
            "type": "object",
            "properties": {
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "financing-offer_internal_core_entity.AssetType": {
            "type": "string",
            "enum": [
                "UNDERLYING",
                "DERIVATIVE"
            ],
            "x-enum-varnames": [
                "AssetTypeUnderlying",
                "AssetTypeDerivative"
            ]
        },
        "financing-offer_internal_core_entity.AwaitingConfirmRequest": {
            "type": "object",
            "properties": {
                "investor": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Investor"
                },
                "latestUpdate": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.OfflineOfferUpdate"
                },
                "loanOffer": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOffer"
                },
                "loanPackageIds": {
                    "type": "string"
                },
                "loanRequest": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequest"
                },
                "symbol": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Symbol"
                }
            }
        },
        "financing-offer_internal_core_entity.CancelledReason": {
            "type": "string",
            "enum": [
                "UNKNOWN",
                "EXPIRED",
                "INVESTOR",
                "ADMIN",
                "ALTERNATIVE_OPTION",
                "HIGH_LOAN_RATE"
            ],
            "x-enum-varnames": [
                "LoanPackageOfferCancelledReasonUnknown",
                "LoanPackageOfferCancelledReasonExpired",
                "LoanPackageOfferCancelledReasonInvestor",
                "LoanPackageOfferCancelledReasonAdmin",
                "LoanPackageOfferCancelledReasonAlternativeOption",
                "LoanPackageOfferCancelledReasonHighLoanRate"
            ]
        },
...
        "financing-offer_internal_core_entity.FlowType": {
            "type": "string",
            "enum": [
                "001",
                "002",
                "ONLINE",
                "OFFLINE"
            ],
            "x-enum-varnames": [
                "FLowTypeDnseOffline",
                "FlowTypeDnseOnline",
                "Online",
                "Offline"
            ]
        },
        "financing-offer_internal_core_entity.Investor": {
            "type": "object",
            "properties": {
                "custodyCode": {
                    "type": "string"
                },
                "investorId": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.InvestorAccount": {
            "type": "object",
            "properties": {
                "accountNo": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "investorId": {
                    "type": "string"
                },
                "marginStatus": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.MarginStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "financing-offer_internal_core_entity.LoanPackageOffer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expiredAt": {
                    "type": "string"
                },
                "flowType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.FlowType"
                },
                "id": {
                    "type": "integer"
                },
                "loanPackageOfferInterests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOfferInterest"
                    }
                },
                "loanPackageRequest": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequest"
                },
                "loanPackageRequestId": {
                    "type": "integer"
                },
                "offeredBy": {
                    "type": "string"
                },
                "symbol": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Symbol"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.LoanPackageOfferInterest": {
            "type": "object",
            "properties": {
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "cancelledAt": {
                    "type": "string"
                },
                "cancelledBy": {
                    "type": "string"
                },
                "cancelledReason": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.CancelledReason"
                },
                "contractSize": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "feeRate": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "initialRate": {
                    "type": "number"
                },
                "interestRate": {
                    "type": "number"
                },
                "limitAmount": {
                    "type": "number"
                },
                "loanContract": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanContract"
                },
                "loanId": {
                    "type": "integer"
                },
                "loanPackageOffer": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOffer"
                },
                "loanPackageOfferId": {
                    "type": "integer"
                },
                "loanRate": {
                    "type": "number"
                },
                "scoreGroupInterestId": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOfferInterestStatus"
                },
                "submissionSheetDetailId": {
                    "type": "integer"
                },
                "term": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "financing-offer_internal_core_entity.LoanPackageRequest": {
            "type": "object",
            "properties": {
                "accountNo": {
                    "type": "string"
                },
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "contractSize": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "guaranteedDuration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "initialRate": {
                    "type": "number"
                },
                "investor": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Investor"
                },
                "investorId": {
                    "type": "string"
                },
                "limitAmount": {
                    "type": "number"
                },
                "loanRate": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequestStatus"
                },
                "symbolId": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequestType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.LoanPackageRequestStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "CONFIRMED"
            ],
            "x-enum-varnames": [
                "LoanPackageRequestStatusPending",
                "LoanPackageRequestStatusConfirmed"
            ]
        },
        "financing-offer_internal_core_entity.LoanPackageRequestType": {
            "type": "string",
            "enum": [
                "FLEXIBLE",
                "GUARANTEED"
            ],
            "x-enum-varnames": [
                "LoanPackageRequestTypeFlexible",
                "LoanPackageRequestTypeGuaranteed"
            ]
        },
...
        "financing-offer_internal_core_entity.MarginStatus": {
            "type": "string",
            "enum": [
                "v2",
                "v3"
            ],
            "x-enum-varnames": [
                "MarginStatusVersion2",
                "MarginStatusVersion3"
            ]
        },
        "financing-offer_internal_core_entity.OfflineOfferUpdate": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "offerId": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.OfflineOfferUpdateStatus"
                }
            }
        },
        "financing-offer_internal_core_entity.OfflineOfferUpdateStatus": {
            "type": "string",
            "enum": [
                "PROCESSING",
                "REJECTED",
                "APPROVED"
            ],
            "x-enum-varnames": [
                "OfflineOfferUpdateStatusProcessing",
                "OfflineOfferUpdateStatusRejected",
                "OfflineOfferUpdateStatusApproved"
            ]
        },
...
        "financing-offer_internal_core_entity.Symbol": {
            "type": "object",
            "properties": {
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastUpdatedBy": {
                    "type": "string"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScore"
                    }
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolStatus"
                },
                "stockExchangeId": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.SymbolScore": {
            "type": "object",
            "properties": {
                "affectedFrom": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "INACTIVE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScoreStatus"
                        }
                    ]
                },
                "symbolId": {
                    "type": "integer"
                },
                "type": {
                    "enum": [
                        "MANUAL",
                        "SYSTEM"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScoreType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.SymbolScoreStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "SymbolScoreStatusActive",
                "SymbolScoreStatusInactive"
            ]
        },
...
        "financing-offer_internal_core_entity.SymbolStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "SymbolStatusActive",
                "SymbolStatusInactive"
            ]
        },
...
        "financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_InvestorAccount": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.InvestorAccount"
                }
            }
        },
...
        "financing-offer_internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_handler.ResponseWithPaging-array_financing-offer_internal_core_entity_AwaitingConfirmRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.AwaitingConfirmRequest"
                    }
                },
                "metaData": {
                    "$ref": "#/definitions/financing-offer_internal_core.PagingMetaData"
                }
            }
        },
...
        "internal_core_investor_account_transport_http.VerifyInvestorAccountMarginStatusRequest": {
            "type": "object",
            "required": [
                "investorId"
            ],
            "properties": {
                "investorId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}